Migrate reused controller stub code into central helper.
