+++ test/integration/exercise4_response_test.rb
@@ -7,29 +7,7 @@ require 'pp'
 require_relative '../examples/exercise4'
 
 module Examples
-  class Exercise4Controller
-    attr_accessor :params
-    attr_accessor :last_render_params
-
-    def initialize
-      self.params = {}
-    end
-
-    def show
-      Flappi.build_and_respond(self)
-    end
-
-    def request
-      OpenStruct.new(query_parameters: params)
-    end
-
-    def respond_to
-      yield JsonFormatter.new
-    end
-
-    def render(params)
-      self.last_render_params = params
-    end
+  class Exercise4Controller < ExampleController
   end
 end
